35/36 tests are passed, problems with GET users and then POST user. I got 400 instead of 200

+++ src/Users/handers/removeUserHandler.handler.ts
@@ -4,18 +4,17 @@ import {queryRepo} from "../../core/repository/data-acsess-present-layer";
 import {usersService} from "../Users-BLL/usersService.BLL";
 import {httpStatus} from "../../core/core-types/http-statuses";
 
-export async function removeUserHandler (req:Request, res:Response){
-    try {
+export async function removeUserHandler (req:Request, res:Response): Promise<void>{
         //получаем id из req.params
         const id = req.params.id;
         //ищем в квериРепо такое или фэйлим
         const foundUser = await queryRepo.findUserByIdOrFail(id);
+        if(foundUser === null){
+            res.sendStatus(httpStatus.NotFound).send({error: "User not found"});
+            return
+        }
         //а потом через БЛЛ и Репозиторий удаляем из БД
         await usersService.removeUserById(foundUser.id)
         //возвращаем статус
         res.sendStatus(httpStatus.NoContent)
-    }
-    catch (e) {
-        errorsHandler(e,res)
-    }
 }
